# services/identity/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: FinCore Identity — Authentication & Access API
  version: "1.0-draft"
  description: >
    The platform's first-party, client-driven identity service (ADR 0018). Every flow is a JSON
    API call; there is no hosted login page. This document is the **authoritative surface** for a
    banking core's authentication needs. Each operation is tagged with `x-status`:

      * **built**   — implemented in this service today (the ADR 0018 swap slice).
      * **planned** — designed and specified here; implementation staged. Where an operation needs
        to deliver a code to a person (email, SMS), it is additionally **delivery-gated**: the
        platform's messaging connector does not exist yet (notification senders are log adapters by
        decision), so the flow's logic is built but the actual send waits on that connector.

    Two surfaces, different callers:
      * **Authentication** (`/v1/auth/**`, `/.well-known/**`) — pre-identity, reached by clients
        through the edge. These are the service's open paths.
      * **Administration** (`/v1/directory/**`, `/v1/roles`, `/v1/permissions`, `/v1/policy`,
        `/v1/audit/**`) — service-facing, driven by Core's admin surface; deny-by-default via a
        bearer token; never routed by the edge; tenants never call it directly.
  contact:
    name: FinCore
  license:
    name: AGPL-3.0-only
servers:
  - url: /
...
paths:
...
  /v1/auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate a staff credential
      description: >
        Verifies username and password for this instance's institution (the tenant is the
        instance, never body-supplied). Returns tokens, or a 202 obligation the client must
        satisfy first — a forced password change, or an MFA challenge when 2FA is active.
      operationId: login
      x-status: built
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated — access and refresh tokens.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '202':
          description: >
            An obligation stands between the credential and tokens. `reason` is
            `PASSWORD_CHANGE_REQUIRED` (a temporary credential) or `MFA_REQUIRED` (2FA active).
            The `actionToken` is single-purpose and short-lived.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionRequired' }
        '401': { $ref: '#/components/responses/AuthFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate a refresh token
      description: >
        Exchanges a refresh token for a fresh access token and a rotated refresh token. Reuse of an
        already-rotated token is treated as theft: the entire family is revoked and the attempt is
        audited. Every refusal is the same opaque `TOKEN_INVALID`.
      operationId: refresh
      x-status: built
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: A fresh access token and a rotated refresh token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/TokenInvalid' }
  /v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke the presented refresh token's family
      operationId: logout
      x-status: built
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LogoutRequest' }
      responses:
        '204': { description: The family is revoked (revoking nothing is not an error). }
  /v1/auth/sessions/revoke-all:
    post:
      tags: [Authentication, Sessions & devices]
      summary: Revoke every session for the authenticated user
      operationId: revokeAllSessions
      x-status: built
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Every refresh family for the bearer is revoked. }
        '401': { $ref: '#/components/responses/TokenInvalid' }
  /v1/auth/token:
    post:
      tags: [Authentication]
      summary: Mint a service token (client-credentials)
      description: >
        A trusted service exchanges its client id and secret for a short-lived service token —
        `azp` set, no tenant claim — exactly what the ledger's caller rules expect.
      operationId: serviceToken
      x-status: built
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ServiceTokenRequest' }
      responses:
        '200':
          description: A service access token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServiceToken' }
        '401': { $ref: '#/components/responses/AuthFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }

  # ============================ Password ============================
  /v1/auth/password:
    post:
      tags: [Password]
      summary: Change a password
      description: >
        Sets a new password — either with the single-purpose action grant from a
        temporary-credential login (forced change), or with the current credential (voluntary
        change). A successful change revokes every session the user holds.
      operationId: changePassword
      x-status: built
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordChangeRequest' }
      responses:
        '204': { description: Password changed; all sessions revoked. }
        '400': { $ref: '#/components/responses/PasswordPolicy' }
        '401': { $ref: '#/components/responses/AuthFailed' }
...
  /v1/auth/mfa:
    get:
      tags: [Multi-factor]
      summary: The caller's enrolled MFA methods and status
      operationId: mfaStatus
      x-status: built
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Enrolled methods and whether MFA is active.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MfaStatus' }
        '401': { $ref: '#/components/responses/TokenInvalid' }
  /v1/auth/mfa/totp/enroll:
    post:
      tags: [Multi-factor]
      summary: Begin TOTP enrolment
      description: >
        Generates a TOTP secret and a set of one-time recovery codes. Returns the secret, an
        `otpauth://` provisioning URI (for an authenticator app QR), and the recovery codes **once**.
        Enrolment is not active until confirmed by `activate`.
      operationId: totpEnroll
      x-status: built
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: The secret, provisioning URI, and recovery codes (shown once).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TotpEnrolment' }
        '401': { $ref: '#/components/responses/TokenInvalid' }
        '409': { description: TOTP is already active for this user. }
  /v1/auth/mfa/totp/activate:
    post:
      tags: [Multi-factor]
      summary: Activate TOTP by proving a first code
      operationId: totpActivate
      x-status: built
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TotpCode' }
      responses:
        '204': { description: TOTP is now active; subsequent logins require it. }
        '401': { $ref: '#/components/responses/AuthFailed' }
  /v1/auth/mfa/verify:
    post:
      tags: [Multi-factor]
      summary: Complete a login by answering the MFA challenge
      description: >
        Presents the `MFA_REQUIRED` action token from `login` together with a current TOTP code (or
        a recovery code), and receives the token pair. Wrong codes refuse with the uniform voice and
        count toward lockout.
      operationId: mfaVerify
      x-status: built
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MfaVerifyRequest' }
      responses:
        '200':
          description: Authenticated — access and refresh tokens.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/AuthFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/auth/mfa/step-up:
    post:
      tags: [Multi-factor]
      summary: Re-assert MFA for a sensitive action (step-up)
      description: >
        A currently-authenticated user re-proves a factor and receives a short-lived elevated token
        carrying `acr` and a fresh `auth_time`. Approval endpoints above a tier threshold can demand
        recent step-up. A flow Keycloak-over-ROPC could never provide (ADR 0018).
      operationId: stepUp
      x-status: built
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TotpCode' }
      responses:
        '200':
          description: A short-lived elevated access token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ElevatedToken' }
        '401': { $ref: '#/components/responses/AuthFailed' }
  /v1/auth/mfa/disable:
    post:
      tags: [Multi-factor]
      summary: Disable a user's MFA (requires step-up)
      operationId: mfaDisable
      x-status: built
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TotpCode' }
      responses:
        '204': { description: MFA disabled; recovery codes invalidated. }
        '401': { $ref: '#/components/responses/AuthFailed' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A staff access token minted by this service (RS256, verifiable via JWKS).
  responses:
    AuthFailed:
      description: >
        Credential refused. One code, no reason, uniform shape and timing for unknown user, wrong
        password, disabled, locked, wrong MFA code, unknown client — distinguishing them is an
        oracle. The true cause is in the audit trail, never on the wire.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: AUTH_FAILED, message: "credential refused", details: {} }
    TokenInvalid:
      description: Token refused — expired, unknown, rotated or revoked, all indistinguishable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: TOKEN_INVALID, message: "token refused", details: {} }
    RateLimited:
      description: Too many attempts. `Retry-After` in seconds. Applied per source and per account.
      headers:
        Retry-After: { schema: { type: integer }, description: Seconds to wait. }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: RATE_LIMITED, message: "slow down", details: {} }
    PasswordPolicy:
      description: The new password was refused by policy.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            code: PASSWORD_POLICY
            reason: TOO_SHORT
            message: "password refused by policy"
            details: { violations: [TOO_SHORT] }
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: ada.admin }
        password: { type: string, format: password }
        clientId: { type: string, example: fincore-web, description: Defaults to fincore-web. }
    TokenPair:
      type: object
      properties:
        accessToken: { type: string, description: RS256 JWT, 5–10 min. }
        refreshToken: { type: string, description: Opaque, rotating, single-use. }
        expiresIn: { type: integer, description: Access token lifetime in seconds. }
    ActionRequired:
      type: object
      properties:
        code: { type: string, enum: [ACTION_REQUIRED] }
        reason: { type: string, enum: [PASSWORD_CHANGE_REQUIRED, MFA_REQUIRED] }
        actionToken: { type: string, description: Single-purpose, short-lived; inert at any other service. }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
        clientId: { type: string }
    LogoutRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    ServiceTokenRequest:
      type: object
      required: [clientId, clientSecret]
      properties:
        clientId: { type: string, example: core }
        clientSecret: { type: string, format: password }
    ServiceToken:
      type: object
      properties:
        accessToken: { type: string }
        expiresIn: { type: integer }
    PasswordChangeRequest:
      type: object
      properties:
        actionToken: { type: string, description: From a forced-change login. Omit for a voluntary change. }
        username: { type: string, description: Required for a voluntary change. }
        currentPassword: { type: string, format: password }
        newPassword: { type: string, format: password }
...
    MfaStatus:
      type: object
      properties:
        active: { type: boolean }
        methods: { type: array, items: { type: string, enum: [TOTP, SMS] } }
        recoveryCodesRemaining: { type: integer }
    TotpEnrolment:
      type: object
      properties:
        secret: { type: string, description: Base32 TOTP secret. }
        otpauthUri: { type: string, example: "otpauth://totp/FinCore:ada.admin?secret=...&issuer=FinCore" }
        recoveryCodes: { type: array, items: { type: string }, description: Shown once; store securely. }
    TotpCode:
      type: object
      required: [code]
      properties:
        code: { type: string, example: "492817" }
    MfaVerifyRequest:
      type: object
      required: [actionToken]
      properties:
        actionToken: { type: string, description: The MFA_REQUIRED action token from login. }
        code: { type: string, description: A TOTP code. }
        recoveryCode: { type: string, description: A recovery code, as an alternative to `code`. }
        clientId: { type: string }
...
    ElevatedToken:
      type: object
      properties:
        accessToken: { type: string, description: Short-lived, carries acr and a fresh auth_time. }
        expiresIn: { type: integer }
...
    Error:
      type: object
      properties:
        code: { type: string }
        reason: { type: string, nullable: true }
        message: { type: string, description: Developer English; never shown to an end user. }
        details: { type: object, additionalProperties: true }